fix(backend,shared): move emit socket service to notification service tm-260
