+++ apps/backend/src/modules/friends/friend.service.ts
@@ -1,8 +1,3 @@
-import {
-	SocketEvent,
-	SocketNamespace,
-	type SocketService,
-} from "~/libs/modules/socket/socket.js";
 import { type FriendRepository } from "~/modules/friends/friend.repository.js";
 import {
 	type NotificationService,
@@ -19,16 +14,13 @@ import {
 class FriendService {
 	private friendRepository: FriendRepository;
 	private notificationService: NotificationService;
-	private socketService: SocketService;
 
 	public constructor(
 		friendRepository: FriendRepository,
 		notificationService: NotificationService,
-		socketService: SocketService,
 	) {
 		this.friendRepository = friendRepository;
 		this.notificationService = notificationService;
-		this.socketService = socketService;
 	}
 
 	public async createSubscription(
@@ -60,17 +52,10 @@ class FriendService {
 			followingUserId,
 		});
 
-		const { notification, receiverId } = await this.notificationService.create({
+		await this.notificationService.create({
 			receiverUserId: followingUserId,
 			type: NotificationType.NEW_FOLLOWER,
 			userId: followerUserId,
-		});
-
-		this.socketService.emitMessage({
-			event: SocketEvent.NEW_NOTIFICATION,
-			payload: notification,
-			receiversIds: [String(notification.userId), String(receiverId)],
-			targetNamespace: SocketNamespace.NOTIFICATIONS,
 		});
 
 		return followingUser.toObject();
