+++ apps/backend/src/modules/friends/friends.ts
@@ -1,5 +1,4 @@
 import { logger } from "~/libs/modules/logger/logger.js";
-import { socketService } from "~/libs/modules/socket/socket.js";
 import { notificationService } from "~/modules/notifications/notifications.js";
 import { UserModel } from "~/modules/users/user.model.js";
 
@@ -8,11 +7,7 @@ import { FriendRepository } from "./friend.repository.js";
 import { FriendService } from "./friend.service.js";
 
 const friendRepository = new FriendRepository(UserModel);
-const friendService = new FriendService(
-	friendRepository,
-	notificationService,
-	socketService,
-);
+const friendService = new FriendService(friendRepository, notificationService);
 const friendController = new FriendController(logger, friendService);
 
 export { friendController };
