+++ apps/backend/src/modules/notifications/libs/types/types.ts
@@ -2,6 +2,5 @@ export {
 	type AllNotificationsResponseDto,
 	type CreateNotificationRequestDto,
 	type NotificationResponseDto,
-	type NotificationWithReceiverIdResponseDto,
 	type UpdateNotificationRequestDto,
 } from "@trackmates/shared";
