+++ apps/backend/src/modules/notifications/notifications.ts
@@ -1,9 +1,14 @@
+import { socketService } from "~/libs/modules/socket/socket.js";
+
 import { NotificationModel } from "./notification.model.js";
 import { NotificationRepository } from "./notification.repository.js";
 import { NotificationService } from "./notification.service.js";
 
 const notificationRepository = new NotificationRepository(NotificationModel);
-const notificationService = new NotificationService(notificationRepository);
+const notificationService = new NotificationService(
+	notificationRepository,
+	socketService,
+);
 
 export { notificationService };
 export { NotificationType } from "./libs/enums/enums.js";
