+++ packages/shared/src/index.ts
@@ -146,7 +146,6 @@ export {
 	type NotificationResponseDto,
 	NotificationStatus,
 	NotificationType,
-	type NotificationWithReceiverIdResponseDto,
 	type UpdateNotificationRequestDto,
 } from "./modules/notifications/notifications.js";
 export {
