--- packages/shared/src/modules/notifications/libs/types/notification-with-receiver-id-response-dto.type.ts
@@ -1,8 +1,0 @@
-import { type NotificationResponseDto } from "./notification-response-dto.type.js";
-
-type NotificationWithReceiverIdResponseDto = {
-	notification: NotificationResponseDto;
-	receiverId: number;
-};
-
-export { type NotificationWithReceiverIdResponseDto };
