+++ packages/shared/src/modules/notifications/libs/types/types.ts
@@ -1,5 +1,4 @@
 export { type AllNotificationsResponseDto } from "./all-notifications-response-dto.type.js";
 export { type CreateNotificationRequestDto } from "./create-notification-request-dto.type.js";
 export { type NotificationResponseDto } from "./notification-response-dto.type.js";
-export { type NotificationWithReceiverIdResponseDto } from "./notification-with-receiver-id-response-dto.type.js";
 export { type UpdateNotificationRequestDto } from "./update-notification-request-dto.type.js";
