+++ packages/shared/src/modules/notifications/notifications.ts
@@ -4,6 +4,5 @@ export {
 	type AllNotificationsResponseDto,
 	type CreateNotificationRequestDto,
 	type NotificationResponseDto,
-	type NotificationWithReceiverIdResponseDto,
 	type UpdateNotificationRequestDto,
 } from "./libs/types/types.js";
